requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.5
scipy>=1.7.3
h5py>=3.7.0
PyYAML>=6.0
matplotlib>=3.5.2
scikit-image>=0.19.2
scikit-learn>=1.0.2  # For t-SNE
tqdm>=4.62.3
lpips>=0.1.4         # For LPIPS metric